Remove CommandParser
Only ButtonClick and KeyboardPress were using the CommandParser. At this
time, it is not necessary to have a dedicated command parser, since commands
are not run very often. Having a command parser makes it possible to queue
commands and print debugging information. However, at this time, that is not
necessary. A command parser may be added back later if needed.